requirements.txt
Pillow
aiogram
numpy